+++ run_test.sh
@@ -1,0 +1,4 @@
+pytest test_textllm.py \
+    --cov \
+    --cov-report term \
+    --cov-report html # --cov-report annotate